+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - [feat] Add laravel 10 project
 - [feat] Add word migration
 - [docs] Update changelog file
+- [feat] Add POST and GET ALL endpoint words